DICT-0011 move and rename dictionary database
