+++ app/src/main/java/ru/easycode/words504/dictionary/data/cache/DictionaryDataBase.kt
@@ -1,14 +1,13 @@
-package ru.easycode.words504
+package ru.easycode.words504.dictionary.data.cache
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import ru.easycode.words504.dictionary.data.cache.WordsDao
 import ru.easycode.words504.dictionary.data.cache.entities.SentenceCache
 import ru.easycode.words504.dictionary.data.cache.entities.WordCache
 import ru.easycode.words504.translate.data.cache.TranslationsDao
 import ru.easycode.words504.translate.data.cache.entities.TranslationCache
 
-interface AppDataBase {
+interface DictionaryDataBase {
 
     fun wordsDao(): WordsDao
     fun translationDao(): TranslationsDao
@@ -17,5 +16,5 @@ interface AppDataBase {
         entities = [SentenceCache::class, WordCache::class, TranslationCache::class],
         version = 1
     )
-    abstract class Base : RoomDatabase(), AppDataBase
+    abstract class Base : RoomDatabase(), DictionaryDataBase
 }
